[fakeintake] 🚚 move fakeintake server implementation to server
