+++ test/fake-intake/.gitignore
@@ -1,1 +1,1 @@
-fake-intake
+build/*
