+++ test/fake-intake/go.mod
@@ -1,6 +1,6 @@
-module github.com/Datadog/datadog-agent/test/fake-intake
+module github.com/Datadog/datadog-agent/test/fakeintake
 
-go 1.19
+go 1.18
 
 require github.com/stretchr/testify v1.8.1
