+++ test/fake-intake/server/server_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package fakeintake
+package server
 
 import (
 	"encoding/json"
@@ -18,9 +18,9 @@ import (
 
 const testFakeIntakePort = 8081
 
-func TestFakeIntake(t *testing.T) {
+func TestServer(t *testing.T) {
 	t.Run("should accept payloads on any route", func(t *testing.T) {
-		fi := NewFakeIntake(testFakeIntakePort)
+		fi := NewServer(testFakeIntakePort)
 		defer fi.server.Close()
 
 		request, err := http.NewRequest(http.MethodPost, "/api/v2/series", strings.NewReader("totoro|5|tag:valid,owner:pducolin"))
@@ -33,7 +33,7 @@ func TestFakeIntake(t *testing.T) {
 	})
 
 	t.Run("should accept GET requests on any other route", func(t *testing.T) {
-		fi := NewFakeIntake(testFakeIntakePort)
+		fi := NewServer(testFakeIntakePort)
 		defer fi.server.Close()
 
 		request, err := http.NewRequest(http.MethodGet, "/api/v1/validate", nil)
@@ -46,7 +46,7 @@ func TestFakeIntake(t *testing.T) {
 	})
 
 	t.Run("should accept GET requests on /fakeintake/payloads route", func(t *testing.T) {
-		fi := NewFakeIntake(testFakeIntakePort)
+		fi := NewServer(testFakeIntakePort)
 		defer fi.server.Close()
 
 		request, err := http.NewRequest(http.MethodGet, "/fakeintake/payloads?endpoint=/foo", nil)
@@ -69,7 +69,7 @@ func TestFakeIntake(t *testing.T) {
 	})
 
 	t.Run("should not accept GET requests on /fakeintake/payloads route without endpoint query parameter", func(t *testing.T) {
-		fi := NewFakeIntake(testFakeIntakePort)
+		fi := NewServer(testFakeIntakePort)
 		defer fi.server.Close()
 
 		request, err := http.NewRequest(http.MethodGet, "/fakeintake/payloads", nil)
@@ -82,7 +82,7 @@ func TestFakeIntake(t *testing.T) {
 	})
 
 	t.Run("should store multiple payloads on any route and return them", func(t *testing.T) {
-		fi := NewFakeIntake(testFakeIntakePort)
+		fi := NewServer(testFakeIntakePort)
 		defer fi.server.Close()
 
 		request, err := http.NewRequest(http.MethodPost, "/api/v2/series", strings.NewReader("totoro|5|tag:valid,owner:pducolin"))
